MAD: Add option for forced sheet update
It's now possible to force the script to download updated data from the
sheet ***now*** through clicking a button in the TamperMonkey context
menu. Sadly, no such option exists in GreaseMonky. There is a polyfill,
but it uses the page's context menu instead of the extension's menu, so
we're not using it in order not to clutter the user's page context menu.