Closes #2
GreaseMonkey doesn't support `@grant window.close`, which causes the
NsDilemmaAutoClose scripts to produce the following error:
`Scripts may not close windows that were not opened by script`

`window.close()` can also be successfully used by a script in another
case:

A browsing context is script-closable if it is an auxiliary browsing
context that was created by a script (as opposed to by an action of the
user), or if it is a top-level browsing context whose session history
contains only one Document.

from: https://www.w3.org/TR/html51/browsers.html#dom-window-close

NsIssueCompactorRand now sets the form target to `_blank`, so that the
browser submits it in a new tab (the new tab thus has only one Document
in its session history, allowing NsDilemmaAutoClose to close the tab).
(Because the form is submitted in a new tab, the answering keybind closes
the current tab in addition to answering the issue.)